Added functionality for calculating gradients of piecewise linear functions on each simplex.
The function 'grad' yields a sparse matrix which maps vertex values to
gradients for each simplex in the mesh (in arbitrary dimensions and
manifolds).
A test case in 2D is included.

Also, let most of the functions in MapToSimp work with matrices instead
of just vectors. This does not affect the old functionality since
Eigen::VectorXd is a wrapper for Eigen::MatrixXd with one column.