Mainly fixed issues with acquiring observation matrices.
* getASimplexForPoint should not return error if one point fail to find
a simplex. This can instead be discovered by the simplex index for that
point being equal to the number of simplices in the mesh. This is fixed
for both mesh, implicitmesh, and hyperRect mesh.

* The area of a (hyper-)parallellogram needs to be divided by the
topological dimensionality in order to acquire the area of the
corresponding simplex. This concerns
"getGradientCahinCoefficientsOfSimplex".

* Get indicator matrix of which nodes that belongs to which simplex.
This concerns function "S2N".

* Boundaries of mesh graph is now also affected by points when acquiring
observation matrix.